Move v1 helper functions into openstack_instances.go
